feat(application): estrutura para consultas de Albums

+++ seletivo-seplag-api/src/main/java/br/com/tcloss/seletivoseplagapi/application/dtos/output/Pagination.java
@@ -1,10 +1,10 @@
 package br.com.tcloss.seletivoseplagapi.application.dtos.output;
 
 public record Pagination(
-    Integer totalPages,
-    Integer currentPage,
-    Integer itemsPerPage,
-    Integer totalIems
+    int totalPages,
+    int currentPage,
+    int itemsPerPage,
+    long totalIems
 ) {
 
 }
